fix: static cast argc to size_t

+++ examples/main.cpp
@@ -14,7 +14,7 @@ int main(int argc, char const *argv[]) {
                .Flg<"disable-content-trust">({.help = "Skip image verification"})
                .Flg<"quiet", "q">({.help = "Supress verbose output"});
 
-  auto const map = parse(p, std::span<char const *>{argv, argc});
+  auto const map = parse(p, std::span<char const *>{argv, static_cast<size_t>(argc)});
   std::print("\nargs map (size {}):\n", map.args.size());
   std::print("name: {}\n", map.get<"name">());
   std::print("platform: {}\n", map.get<"platform">());
